move devtools extensions tests
