+++ spec/api-browser-window-spec.js
@@ -112,234 +112,6 @@ describe('BrowserWindow module', () => {
 
   afterEach(closeTheWindow)
 
-  describe('window.webContents.send(channel, args...)', () => {
-    it('throws an error when the channel is missing', () => {
-      expect(() => {
-        w.webContents.send()
-      }).to.throw('Missing required channel argument')
-
-      expect(() => {
-        w.webContents.send(null)
-      }).to.throw('Missing required channel argument')
-    })
-  })
-
-  describe('window.getNativeWindowHandle()', () => {
-    before(function () {
-      if (!nativeModulesEnabled) {
-        this.skip()
-      }
-    })
-
-    it('returns valid handle', () => {
-      // The module's source code is hosted at
-      // https://github.com/electron/node-is-valid-window
-      const isValidWindow = remote.require('is-valid-window')
-      expect(isValidWindow(w.getNativeWindowHandle())).to.be.true()
-    })
-  })
-
-  describe('extensions and dev tools extensions', () => {
-    let showPanelTimeoutId
-
-    const showLastDevToolsPanel = () => {
-      w.webContents.once('devtools-opened', () => {
-        const show = () => {
-          if (w == null || w.isDestroyed()) return
-          const { devToolsWebContents } = w
-          if (devToolsWebContents == null || devToolsWebContents.isDestroyed()) {
-            return
-          }
-
-          const showLastPanel = () => {
-            const lastPanelId = UI.inspectorView._tabbedPane._tabs.peekLast().id
-            UI.inspectorView.showPanel(lastPanelId)
-          }
-          devToolsWebContents.executeJavaScript(`(${showLastPanel})()`, false).then(() => {
-            showPanelTimeoutId = setTimeout(show, 100)
-          })
-        }
-        showPanelTimeoutId = setTimeout(show, 100)
-      })
-    }
-
-    afterEach(() => {
-      clearTimeout(showPanelTimeoutId)
-    })
-
-    describe('BrowserWindow.addDevToolsExtension', () => {
-      describe('for invalid extensions', () => {
-        it('throws errors for missing manifest.json files', () => {
-          const nonexistentExtensionPath = path.join(__dirname, 'does-not-exist')
-          expect(() => {
-            BrowserWindow.addDevToolsExtension(nonexistentExtensionPath)
-          }).to.throw(/ENOENT: no such file or directory/)
-        })
-
-        it('throws errors for invalid manifest.json files', () => {
-          const badManifestExtensionPath = path.join(__dirname, 'fixtures', 'devtools-extensions', 'bad-manifest')
-          expect(() => {
-            BrowserWindow.addDevToolsExtension(badManifestExtensionPath)
-          }).to.throw(/Unexpected token }/)
-        })
-      })
-
-      describe('for a valid extension', () => {
-        const extensionName = 'foo'
-
-        const removeExtension = () => {
-          BrowserWindow.removeDevToolsExtension('foo')
-          expect(BrowserWindow.getDevToolsExtensions()).to.not.have.a.property(extensionName)
-        }
-
-        const addExtension = () => {
-          const extensionPath = path.join(__dirname, 'fixtures', 'devtools-extensions', 'foo')
-          BrowserWindow.addDevToolsExtension(extensionPath)
-          expect(BrowserWindow.getDevToolsExtensions()).to.have.a.property(extensionName)
-
-          showLastDevToolsPanel()
-
-          w.loadURL('about:blank')
-        }
-
-        // After* hooks won't be called if a test fail.
-        // So let's make a clean-up in the before hook.
-        beforeEach(removeExtension)
-
-        describe('when the devtools is docked', () => {
-          beforeEach(function (done) {
-            addExtension()
-            w.webContents.openDevTools({ mode: 'bottom' })
-            ipcMain.once('answer', (event, message) => {
-              this.message = message
-              done()
-            })
-          })
-
-          describe('created extension info', function () {
-            it('has proper "runtimeId"', function () {
-              expect(this.message).to.have.own.property('runtimeId')
-              expect(this.message.runtimeId).to.equal(extensionName)
-            })
-            it('has "tabId" matching webContents id', function () {
-              expect(this.message).to.have.own.property('tabId')
-              expect(this.message.tabId).to.equal(w.webContents.id)
-            })
-            it('has "i18nString" with proper contents', function () {
-              expect(this.message).to.have.own.property('i18nString')
-              expect(this.message.i18nString).to.equal('foo - bar (baz)')
-            })
-            it('has "storageItems" with proper contents', function () {
-              expect(this.message).to.have.own.property('storageItems')
-              expect(this.message.storageItems).to.deep.equal({
-                local: {
-                  set: { hello: 'world', world: 'hello' },
-                  remove: { world: 'hello' },
-                  clear: {}
-                },
-                sync: {
-                  set: { foo: 'bar', bar: 'foo' },
-                  remove: { foo: 'bar' },
-                  clear: {}
-                }
-              })
-            })
-          })
-        })
-
-        describe('when the devtools is undocked', () => {
-          beforeEach(function (done) {
-            addExtension()
-            w.webContents.openDevTools({ mode: 'undocked' })
-            ipcMain.once('answer', (event, message, extensionId) => {
-              this.message = message
-              done()
-            })
-          })
-
-          describe('created extension info', function () {
-            it('has proper "runtimeId"', function () {
-              expect(this.message).to.have.own.property('runtimeId')
-              expect(this.message.runtimeId).to.equal(extensionName)
-            })
-            it('has "tabId" matching webContents id', function () {
-              expect(this.message).to.have.own.property('tabId')
-              expect(this.message.tabId).to.equal(w.webContents.id)
-            })
-          })
-        })
-      })
-    })
-
-    it('works when used with partitions', (done) => {
-      if (w != null) {
-        w.destroy()
-      }
-      w = new BrowserWindow({
-        show: false,
-        webPreferences: {
-          nodeIntegration: true,
-          partition: 'temp'
-        }
-      })
-
-      const extensionPath = path.join(__dirname, 'fixtures', 'devtools-extensions', 'foo')
-      BrowserWindow.removeDevToolsExtension('foo')
-      BrowserWindow.addDevToolsExtension(extensionPath)
-
-      showLastDevToolsPanel()
-
-      ipcMain.once('answer', function (event, message) {
-        expect(message.runtimeId).to.equal('foo')
-        done()
-      })
-
-      w.loadURL('about:blank')
-      w.webContents.openDevTools({ mode: 'bottom' })
-    })
-
-    it('serializes the registered extensions on quit', () => {
-      const extensionName = 'foo'
-      const extensionPath = path.join(__dirname, 'fixtures', 'devtools-extensions', extensionName)
-      const serializedPath = path.join(app.getPath('userData'), 'DevTools Extensions')
-
-      BrowserWindow.addDevToolsExtension(extensionPath)
-      app.emit('will-quit')
-      expect(JSON.parse(fs.readFileSync(serializedPath))).to.deep.equal([extensionPath])
-
-      BrowserWindow.removeDevToolsExtension(extensionName)
-      app.emit('will-quit')
-      expect(fs.existsSync(serializedPath)).to.be.false()
-    })
-
-    describe('BrowserWindow.addExtension', () => {
-      beforeEach(() => {
-        BrowserWindow.removeExtension('foo')
-        expect(BrowserWindow.getExtensions()).to.not.have.property('foo')
-
-        const extensionPath = path.join(__dirname, 'fixtures', 'devtools-extensions', 'foo')
-        BrowserWindow.addExtension(extensionPath)
-        expect(BrowserWindow.getExtensions()).to.have.property('foo')
-
-        showLastDevToolsPanel()
-
-        w.loadURL('about:blank')
-      })
-
-      it('throws errors for missing manifest.json files', () => {
-        expect(() => {
-          BrowserWindow.addExtension(path.join(__dirname, 'does-not-exist'))
-        }).to.throw('ENOENT: no such file or directory')
-      })
-
-      it('throws errors for invalid manifest.json files', () => {
-        expect(() => {
-          BrowserWindow.addExtension(path.join(__dirname, 'fixtures', 'devtools-extensions', 'bad-manifest'))
-        }).to.throw('Unexpected token }')
-      })
-    })
-  })
-
   describe('window.webContents.executeJavaScript', () => {
     const expected = 'hello, world!'
     const expectedErrorMsg = 'woops!'
